% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\docType{data}
\name{CountryPopulation}
\alias{CountryPopulation}
\title{The population for a given country.}
\format{
a function
\describe{
\item{Year}{The year that was recorded}
\item{Population}{The population of that year}
\item{Country}{The country}
}
}
\usage{
CountryPopulation
}
\description{
The population for a given country.
}
\keyword{datasets}
