% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldCup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{The attendance details of the world cups}
\format{
a data frame with 22 observations
\describe{
\item{WorldCup}{The year and country}
\item{Matches}{The number of matches in that competition}
\item{Total Attendance}{How many people attended the game}
\item{Average Attendance}{The average of how many people attended the game}
}
}
\usage{
World_Cup
}
\description{
The attendance details of the world cups
}
\keyword{datasets}
