% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{The population over time for the UN.}
\format{
a data frame with 235 observations
\describe{
\item{Year}{The year that was recorded}
\item{Population}{The population of that year}
\item{Country}{The country}
}
}
\usage{
WorldPopulation
}
\description{
The population over time for the UN.
}
\keyword{datasets}
